+++ include/core/Space.hpp
@@ -3,7 +3,9 @@
 
 #include <map>
 #include <string>
+
 #include "core/Camera.hpp"
+#include "core/Window.hpp"
 
 // a utility to group several objects, similar to a layer.
 
@@ -12,12 +14,22 @@
 // TODO: implement Draw().
 
 // TODO(maybe):
-// Add a Quadtree.
-// Use a map to save Object id-s.
+// Add a Quadtree ??
 
 namespace tnt
 {
     class Object;
+    class Window;
+
+    namespace detail
+    {
+        auto should_draw = [](tnt::Rectangle const &bounds, tnt::Vector const &pos, float w, float h) -> bool {
+            return (bounds.Contains(pos) ||
+                    bounds.Contains(pos + tnt::Vector{w, 0}) ||
+                    bounds.Contains(pos + tnt::Vector{0, h}) ||
+                    bounds.Contains(pos + tnt::Vector{w, h}));
+        };
+    } // namespace detail
 
     class Space
     {
@@ -28,15 +40,31 @@ namespace tnt
         Object *getObject(std::string_view id) const;
         void removeObject(std::string_view id) noexcept;
 
-        // TODO: unimplemented
         template <camera C>
-        void Draw(Window const *win, C const &cam) noexcept;
+        void Draw(Window const *win, C const &cam) noexcept
+        {
+            for (auto const &it : objects)
+                if (it.second->isActive() &&
+                    it.second->has<SpriteComponent>())
+                {
+                    Sprite *spr{dynamic_cast<Sprite *>(it.second)};
+                    SpriteComponent *sprite{spr->getSprite()};
 
-        void Update(long long time_);
+                    const int w{sprite->getWidth()};
+                    const int h{sprite->getHeight()};
+
+                    if (tnt::Vector pos{spr->getPosition()};
+                        detail::should_draw(cam.Bounds(), pos, static_cast<float>(w), static_cast<float>(h)))
+                        win->Draw(spr,
+                                  SDL_Rect{static_cast<int>(pos.x), static_cast<int>(pos.y), w, h}, cam);
+                }
+        }
+
+        void Update(long long time_) noexcept;
 
     protected:
         std::map<std::string, Object *> objects;
     };
 } // namespace tnt
 
-#endif //! TNT_SPACE_HPP
+#endif //!TNT_SPACE_HPP