+++ scripts/player.lua
@@ -1,0 +1,14 @@
+-- -- temporary workaround for handling keys
+-- local RIGHT = 79
+-- local LEFT = 80
+-- local DOWN = 81
+-- local UP = 82
+
+-- function update(time_)
+--     if input.key_down(RIGHT) then print "right" end
+--     if input.key_down(LEFT) then print "left" end
+--     if input.key_down(DOWN) then print "down" end
+--     if input.key_down(UP) then print "up" end
+
+--     print(1000/time_, " fps\n")
+-- end