+++ src/core/Space.cpp
@@ -1,6 +1,8 @@
 // This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
+#include <SDL2/SDL.h>
+
 #include "core/Space.hpp"
 #include "ecs/Object.hpp"
 #include "core/Camera.hpp"
@@ -12,7 +14,7 @@ void tnt::Space::addObject(std::string_view id, tnt::Object *obj)
     objects.insert(std::make_pair(id, obj));
 }
 
-tnt::Object *tnt::Space::getObject(std::string_view id) const { return objects.at(id.data()); }
+tnt::Object *tnt::Space::getObject(std::string_view id) const { return objects.find(id.data())->second; }
 
 void tnt::Space::removeObject(std::string_view id) noexcept
 {
@@ -20,27 +22,7 @@ void tnt::Space::removeObject(std::string_view id) noexcept
         it = objects.erase(it);
 }
 
-template <typename T>
-concept drawable_pair = requires(T const &t)
-{
-    {
-        t.second
-    }
-    ->tnt::drawable;
-};
-
-template <tnt::camera C>
-void tnt::Space::Draw(const tnt::Window *win, const C &cam) noexcept
-{
-    for (drawable_pair auto const &it : objects)
-        if (it.second->isActive())
-            if (Rectangle bounds{cam.Bounds()};
-                bounds.Contains(it.second->getPosition()) ||
-                bounds.Contains(it.second->getPosition() + {}))
-                it.second->getSprite()->Draw(win, cam->getBounds());
-}
-
-void tnt::Space::Update(long long time_)
+void tnt::Space::Update(long long time_) noexcept
 {
     for (auto const &it : objects)
         if (it.second->isActive())