+++ test/ecs/main.cpp
@@ -1,3 +1,4 @@
+#include <sol/sol.hpp>
 #include "ecs/Object.hpp"
 
 class Test : public tnt::Component
@@ -16,30 +17,38 @@ private:
 class Player : public tnt::Object
 {
 public:
-    explicit Player(int value) : test{add<Test>(value)} {}
-    ~Player() noexcept { clear(); }
+    explicit Player(int value, std::string_view filename) noexcept
+        : test{add<Test>(value)}, script{add<tnt::Script>(filename)} {}
+
+    ~Player() noexcept
+    {
+        clear();
+        if (script != nullptr)
+            tnt::logger::info("Script just leaked!!");
+    }
 
     void setTest(Test *test_) noexcept { test = test_; }
     Test *getTest() const noexcept { return test; }
 
-    virtual void Update([[maybe_unused]] long long time_) noexcept
+    virtual void Update(long long time_) noexcept
     {
         int val{test->getValue()};
         test->setValue(++val);
         tnt::logger::info("Value: {}", test->getValue());
+        script->Update(time_);
     }
 
 private:
     Test *test;
+    tnt::Script *script;
 };
 
 int main()
 {
-    Player *player{new Player{10}};
+    Player player{10, "player.lua"};
 
     for (int num{0}; num < 1000; ++num)
-        player->Update(10);
+        player.Update(num);
 
-    delete player;
     return 0;
 }