+++ test/ecs/player.lua
@@ -1,0 +1,3 @@
+function update(time_)
+    print(time_, '\n')
+end